BOT/requirements.txt
aiogram==3.17.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10